// src/providers/providers.enum.ts
export enum PROVIDERS {
    DATABASE_CONNECTION = 'DATABASE_CONNECTION',

    MAIN_TICKETS_REPOSITORY = 'MAIN_TICKETS_REPOSITORY',
    FLIGHT_TICKETS_REPOSITORY = 'FLIGHT_TICKETS_REPOSITORY',
    RAILWAY_TICKETS_REPOSITORY = 'RAILWAY_TICKETS_REPOSITORY',
    BUS_TICKETS_REPOSITORY = 'BUS_TICKETS_REPOSITORY',
    ORDERED_TICKETS_REPOSITORY = 'ORDERED_TICKETS_REPOSITORY',

    MAIN_STAYS_REPOSITORY = 'MAIN_STAYS_REPOSITORY',
    HOTELS_REPOSITORY = 'HOTELS_REPOSITORY',
    ROOMS_REPOSITORY = 'ROOMS_REPOSITORY',
    ORDERED_STAYS_REPOSITORY = 'ORDERED_STAYS_REPOSITORY',

    USERS_REPOSITORY = 'USERS_REPOSITORY',

    CARS_REPOSITORY = 'CARS_REPOSITORY',
    RENTED_CARS_REPOSITORY = 'RENTED_CARS_REPOSITORY',

    ENTERTAINMENT_REPOSITORY = 'ENTERTAINMENT_REPOSITORY',
    ORDERED_ENTERTAINMENT_REPOSITORY = 'ORDERED_ENTERTAINMENT_REPOSITORY',

    CITIES_REPOSITORY = 'CITIES_REPOSITORY',
    COUNTRIES_REPOSITORY = 'COUNTRIES_REPOSITORY',
    COMPANIES_REPOSITRY = 'COMPANIES_REPOSITRY',
    IMAGES_REPOSITORY = 'IMAGES_REPOSITORY',
    AMENITIES_REPOSITORY = 'AMENITIES_REPOSITORY',
    
}